Add AppointmentNotFoundException, ErrorResponse, rename exception directory
