+++ src/main/java/kma/health/app/kma_health/exception/AppointmentNotFoundException.java
@@ -1,0 +1,11 @@
+package kma.health.app.kma_health.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class AppointmentNotFoundException extends RuntimeException {
+    public AppointmentNotFoundException(String message) {
+        super(message);
+    }
+}
