+++ src/main/java/kma/health/app/kma_health/exception/CoordinatesNotFoundException.java
@@ -1,4 +1,4 @@
-package kma.health.app.kma_health.exceptions;
+package kma.health.app.kma_health.exception;
 
 public class CoordinatesNotFoundException extends RuntimeException {
     public CoordinatesNotFoundException(String message) {
