+++ src/main/java/kma/health/app/kma_health/exception/ErrorResponse.java
@@ -1,0 +1,21 @@
+package kma.health.app.kma_health.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ErrorResponse {
+
+    private int statusCode;
+    private String message;
+
+    public ErrorResponse(String message)
+    {
+        super();
+        this.message = message;
+    }
+}
